# Implement Map.prototype[ @@toStringTag ]