--- a/boa/src/builtins/map/mod.rs
+++ b/boa/src/builtins/map/mod.rs
@@ -50,7 +50,7 @@ impl BuiltIn for Map {
             )
             .property(
                 to_string_tag,
-                "Math",
+                "Map",
                 Attribute::READONLY | Attribute::NON_ENUMERABLE | Attribute::CONFIGURABLE,
             )
             .property(