This Implements `Map.prototype[ @@toStringTag ] `

This is based on #1225

resolves #1229